Rename package imports after package movement
