+++ lyrebird/src/main/java/moe/lyrebird/view/screens/credits/CreditsScreenController.java
@@ -23,8 +23,8 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 import moe.tristan.easyfxml.model.components.listview.ComponentListViewFxmlController;
-import moe.lyrebird.model.credits.CreditedWork;
 import moe.lyrebird.model.credits.CreditsService;
+import moe.lyrebird.model.credits.objects.CreditedWork;
 import moe.lyrebird.view.components.cells.CreditsCell;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
